moved lengths file write, cast bug fix

+++ include/mem_finder.hpp
@@ -211,13 +211,12 @@ private:
 
     inline bool check_doc_range(size_t start, size_t end) 
     {
-        
         std::unordered_set<size_t> seen;
         size_t unique = 0;
         size_t iterations = end - start + 1;
         size_t idx = 0;
         std::deque<size_t>::iterator it = da_buffer.begin() + (start - buffer_start);
-        uint8_t cur_char = *it;
+        size_t cur_char = *it;
         while (idx < iterations) {
             if (!seen.count(*it)) {
                 unique++;
